# docs/openapi.yaml
openapi: 3.0.3
info:
  title: go-tari-ootle-explorer API
  description: >-
    Read-only JSON API over this repo's four Ootle (L2) explorer data surfaces:
    indexer-backed blocks, the validator roster, L1->L2 burn-claim tracking, and the
    community template registry mirror. Every route in this namespace is served by
    `internal/server/api.go` and is strictly read-only against this repo's own
    Postgres store - it never talks to `tari_indexer`/`tari_validator_node` directly
    (see that file's own doc comment).
  version: "0.1.0"
  contact:
    name: go-tari-ootle-explorer
    url: https://github.com/Snipa22/go-tari-ootle-explorer
servers:
  - url: https://ootle-testnet.supportxtm.com
    description: Public testnet deployment
  - url: http://localhost:8080
    description: Local development (cmd/server default listen address)
...
paths:
...
  /api/validators:
    get:
      operationId: listValidators
      tags: [validators]
      summary: List the full validator roster
      description: >-
        Returns every row in the `validators` table - unpaginated, matching the
        HTML `/validators` route's own no-pagination behavior (a real-world
        validator roster is small enough for v1's purposes).
      responses:
        "200":
          description: >-
            An array of validators. An empty result is always marshaled as a
            JSON array `[]`, never `null`.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ValidatorRow"
        "500":
          description: Store/database failure.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /api/burn-claims:
    get:
      operationId: listBurnClaims
      tags: [burn-claims]
      summary: List L1->L2 burn-claim tracking rows
      description: >-
        Returns `burn_claims` rows, most-recently-mined first (ORDER BY
        burn_height DESC), honoring the same `?status=` filter semantics as the
        HTML `/burn-claims` route.
      parameters:
        - name: status
          in: query
          required: false
          description: >-
            Filters to exactly one status. Omitted/empty returns every status.
            An unrecognized value is passed straight through to the query
            (which will simply match zero rows, since the table's own CHECK
            constraint limits real rows to the three known values) rather than
            rejected with a 400.
          schema:
            type: string
            enum: [pending, claimed, stuck]
      responses:
        "200":
          description: >-
            An array of burn claims. An empty result is always marshaled as a
            JSON array `[]`, never `null`.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BurnClaim"
        "500":
          description: Store/database failure.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
    ApiError:
      type: object
      description: >-
        Shared error-response shape for every /api/* route's 400/500 responses -
        this repo returns valid JSON even on error, never a plain-text body.
      required: [error]
      properties:
        error:
          type: string
          example: failed to load blocks
...
    ValidatorRow:
      type: object
      description: Mirrors `db.ValidatorRow` - the full validators table roster shape.
      required:
        - public_key
        - last_seen_epoch
        - shard_group_start
        - shard_group_end_inclusive
        - consensus_status
        - last_checked_at
      properties:
        public_key:
          type: string
        last_seen_epoch:
          type: integer
          format: int64
          minimum: 0
        shard_group_start:
          type: integer
          format: int32
          minimum: 0
        shard_group_end_inclusive:
          type: integer
          format: int32
          minimum: 0
        consensus_status:
          type: string
          nullable: true
          description: >-
            From `tari_validator_node`'s JSON-RPC get_consensus_status (e.g.
            "Running", "Syncing"), populated by internal/vnhealth. Null if that
            validator has never been checked by vnhealth.
        last_checked_at:
          type: string
          format: date-time
    BurnClaim:
      type: object
      description: Mirrors `db.BurnClaim` - a single burn_claims tracking row.
      required:
        - l1_burn_tx_hash
        - commitment
        - claim_public_key
        - burn_height
        - claim_tx_id
        - claimed_at
        - status
      properties:
        l1_burn_tx_hash:
          type: string
          description: The L1 burn kernel's hash.
        commitment:
          type: string
        claim_public_key:
          type: string
          nullable: true
          description: >-
            Always null coming from internal/burnclaim's L1 scanner - see
            db.BurnClaim's own doc comment.
        burn_height:
          type: integer
          format: int64
          minimum: 0
        claim_tx_id:
          type: string
          nullable: true
          description: >-
            Always null even once status='claimed' - the tombstone substate
            lookup this repo uses to detect a claim doesn't expose an L2
            transaction id.
        claimed_at:
          type: string
          format: date-time
          nullable: true
          description: >-
            This repo's own OBSERVATION time of the claim, not a real upstream
            claim timestamp - null until status='claimed'.
        status:
          type: string
          enum: [pending, claimed, stuck]
...